Rename 'censor' to 'redact', redact UUIDs, test
It's redacting, not strictly censoring, so call it redacting.  Change
the redaction indicator, which is not strictly part of our API
guarantee.

Match on UUID path components and redact those too, since that's what
Heroku are actually using for authentication.

Our First Test!  Table-driven testing of the redaction.

Fix a couple of typos in docs.